Improved command buffer usage efficiency.
Directly pass around the command buffer as the handle so it can be casted straight into a VkCommandBuffer at the usage site instead of having to do an array access to find it. The array would likely be in cache but this is just marginally more efficient. However, we need to do a table lookup to get the command buffer index to release the pool when submitting.

It also means that we don't need to lock the live command buffer array except during start/end recording when I implement multi thread command recording.